simple_trading_bot/requirements.txt
appnope==0.1.2
asyncio==3.4.3
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
enum34==1.1.10
forex-python==1.5
freezegun==1.1.0
idna==2.10
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
lxml==4.6.2
more-itertools==8.6.0
multitasking==0.0.9
numpy==1.19.5
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
Pygments==2.7.4
pyRofex==0.3.1
python-dateutil==2.8.1
pytz==2020.5
Quandl==3.6.0
requests==2.25.1
simplejson==3.17.2
six==1.15.0
tabulate==0.8.7
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
websocket-client==0.57.0
yfinance==0.1.55
